Large Change. See Below:
Introduced a validation suite as per the reference lists
according to Martin Porter's site.

The stemmer.js prior to this commit was neither compliant
with the 1980 nor the Stemmer2 corpi.

As of this commit, it has been split into two files,

 * PorterStemmer1980, corresponding to the original 1980 paper

 * PorterStemmer2, corresponding to the July 2005 reference

Additionally, a validation suite for both stemmers have been
added so that they can be verified as accurate.

The code itself has been cleaned up and a debug call has been
added, see the readme on the site.

As of this commit, PorterStemmer2 is inaccurate, only achieving
a 95% completion on the corpus.  I will get back to this soon
and bring it up to 100%.

Additionall, a licensing term has been introduced, the most
permissive and widely accepted according to a wikipedia comparison
chart.